--- lms/djangoapps/labster_course_license/__init__.py
@@ -1,4 +1,0 @@
-"""
-Labster course license app.
-"""
-from . import signals
