+++ openedx/core/djangoapps/labster/course/signals.py
@@ -14,3 +14,15 @@ def _listen_for_course_delete(sender, course_key, **kwargs):  # pylint: disable=
     # Import tasks here to avoid a circular import.
     from .tasks import course_delete
     course_delete.delay(unicode(course_key))
+
+
+@receiver(SignalHandler.course_published)
+def on_course_publish(sender, course_key, **kwargs):  # pylint: disable=unused-argument
+    """
+    Will receive a delegated 'course_published' signal
+    and kick off a celery task to update the course access structure.
+    """
+    # prevent from circular import
+    from .tasks import update_course_access
+
+    update_course_access.delay(unicode(course_key))
